Don't provide a default value for as-html-id.
See https://github.com/Patternslib/Patterns/issues/149

Also, the camelcasing from parser.js is wrong in this case (and apparently
the feature is removed in the syntax-cleanup branch).

For now, we have to however somehow deal with the fact that "asHtmlId" is
returned when jquery.autoSuggest expects "asHtmlID".

As soon as we start using the syntax-cleanup branch, the above will be obsolete
again, but I assume quite a few changes will be required to this pattern as
well.