Don't close the tooltip when an injection happens.
Couldn't find a reason or any tests to explain the usecase where this might be
necessary.

If there is such a usecase, then we need its particular details, but listening
to pat-inject in general is much too broad.

I've also added some more tests, including a regression tests to ensure this
doesn't get blindly readded again.