Bugfix. Handle case where container event.data is undefined.
I couldn't find a smoking gun for why this is.
It happens on IE when the window is resized and I originally though it had to
do with utils.debounce, but didn't find a bug.

The fact that it's very difficult to debug stuff in IE didn't help.

If we move to closured patterns (like in mockup), then we would avoid a whole
class of bugs similar to this one, because you can just assign whatever you
need to keep track of on the pattern.

I.e. this.$container = $(this);

For each instance of the pattern that would be unique. Currently we have global
scope for a pattern and therefore have to do tricks like settting event.data.